Fix navbar plugin to avoid mandatory white pages for ranks
Previously the "hébergement" section has an empty introduction page showing up due to having an empty index.md page to specify folder title and rank in the navbar.
Now index.md are not displayed if not visible, but their enclosing folder is displayed as soon as at least one of their children is visible.